default cell type is unsigned

+++ kernel/c/sable.c
@@ -5,7 +5,7 @@
 #include <sys/mman.h>
 #include <unistd.h>
 
-typedef unsigned long TYPE;
+typedef unsigned TYPE;
 
 static TYPE *TABLE = NULL;
